+++ sdk/core/Azure.Core/src/ResponseOfT.cs
@@ -19,15 +19,10 @@ namespace Azure
 #pragma warning restore SA1649 // File name should match first type name
     {
         /// <summary>
-        /// Returns the HTTP response returned by the service.
-        /// </summary>
-        /// <returns>The HTTP response returned by the service.</returns>
-        public abstract Response GetRawResponse();
-
-        /// <summary>
         /// Gets the value returned by the service.
         /// </summary>
-        public new abstract T Value { get; }
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public override bool HasValue => true;
 
         /// <summary>
         /// Returns the value of this <see cref="Response{T}"/> object.
@@ -52,11 +47,5 @@ namespace Azure
         /// <inheritdoc />
         [EditorBrowsable(EditorBrowsableState.Never)]
         public override int GetHashCode() => base.GetHashCode();
-
-        /// <inheritdoc />
-        public override string ToString()
-        {
-            return $"Status: {GetRawResponse().Status}, Value: {Value}";
-        }
     }
 }